add: 메인 화면구현, 글 작성, 수정, 첨부파일, 댓글쓰기 기능구현

+++ src/main/java/com/gyemoim/domain/board/BoardVO.java
@@ -8,7 +8,7 @@ import java.util.Date;
 public class BoardVO {
 
   private int bid;
-  private int uNo;
+  private Long uNo;
   private String type;
   private String name;
   private String title;
